Add tick and shadow to chosen answer
