+++ frontend/src/pages/ExamPage.js
@@ -27,6 +27,9 @@ const ExamPage = (props) => {
 	
 	useEffect(() => {
 		examInfo(props.match.params.id, (res) => {
+			res.data.questions.forEach(question => {
+				question.answer = shuffle(question.answer)
+			});
 			setExam(res.data);
 			setLoading(false);
 		})
@@ -54,6 +57,24 @@ const ExamPage = (props) => {
 			</Grid>
 		</div>
 	)
+}
+
+// used to shuffle the answers in the question
+const shuffle = (array) => {
+	let currentIndex = array.length,  randomIndex;
+
+	// While there remain elements to shuffle...
+	while (currentIndex != 0) {
+		// Pick a remaining element...
+		randomIndex = Math.floor(Math.random() * currentIndex);
+		currentIndex--;
+
+		// And swap it with the current element.
+		[array[currentIndex], array[randomIndex]] = [
+		array[randomIndex], array[currentIndex]];
+	}
+
+	return array;
 }
 
 export default ExamPage;
